legacy/src/ultravox/pyproject.toml
[tool.poetry]
name = "ultravox"
version = "0.1.0"
description = "Training engine for E2E voice models"
authors = ["Farzad Abdolhosseini <farzad@fixie.ai>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.9,<3.9.7 || >3.9.7,<4.0"
torch = "^2.1.2"
torchaudio = "^2.1.2"
librosa = "^0.10.1"
transformers = "4.37.2"
datasets = "^2.16.1"
deepspeed = "^0.12.6"
peft = "^0.7.1"
wandb = "^0.16.1"
clearml = "^1.13.2"
pyrallis = "^0.3.1"
truecase = "^0.0.14"
tensorboardx = "^2.6.2.2"
evaluate = "^0.4.1"
azure-keyvault-secrets = "^4.7.0"
azure-identity = "^1.15.0"
azureml-core = "^1.54.0.post1"
bitsandbytes = "^0.42.0"
jiwer = "^3.0.3"
streamlit = "^1.30.0"
gradio = "^4.16.0"
fairseq = "^0.12.2"
sounddevice = "^0.4.6"


[tool.poetry.group.dev.dependencies]
pytest = "^7.4.4"
black = "^23.12.1"
isort = "^5.13.2"
autoflake = "^2.2.1"
deptry = "^0.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
